pyproject.toml
[project]
name = "langchain-learning"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Aryan Iyappan", email = "aryaniyappan2006@gmail.com"},
]
dependencies = [
    "langchain>=0.2.6",
    "langchain-huggingface>=0.0.3",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.5.0",
    "mypy>=1.10.1",
    "black>=24.4.2",
]
